+++ attnroute/__init__.py
@@ -17,7 +17,7 @@ Quick start:
     attnroute status
 """
 
-__version__ = "0.5.6"
+__version__ = "0.5.7"
 __author__ = "jeranaias"
 __all__ = [
     "__version__",